requirements.txt
flask[async]
flask-mysql
python-dotenv